Reflavor drinking amounts, pillows, and maybe ashtrays
- Change the clear cut amounts that you 'transfer liquid' from cups to be something like sips, tastes, gulps, etc.

- Todo: add pillows, the files added but not ticked on yet

- Ashtrays todo: sprites exist but no code is written yet.

Placing cig on ashtray should mark cig as yours, clicking it again picks it up.  Should make a sprite for cigs in the ashtray that aren't just burnt out.  Picking up a cig that isn't marked as yours should clearly state 'so and so picks up the whatever left by the whoever' or something.